feat: add version flags and unify version lookup
